+++ src/sample/Payment.java
@@ -26,18 +26,16 @@ public class Payment extends Transaction {
      */
     public void create() {
         try {
-            Statement stmt = CryptoBudgetDatabase.connection.createStatement();
             String savePayment = "INSERT INTO PAYMENT (USERID, AMOUNT, DATE, ENDDATE, CURRENCYTYPE, FREQUENCY, " +
                     "OTHERPARTY) VALUES (?, ?, ?, ?, ?, ?, ?)";
-
             PreparedStatement prep = CryptoBudgetDatabase.connection.prepareStatement(savePayment);
-            prep.setInt(1, this.userId);
-            prep.setDouble(2, this.amount);
-            prep.setInt(3, this.date);
-            prep.setInt(4, this.endDate);
-            prep.setInt(5, this.currencyType);
-            prep.setInt(6, this.frequency);
-            prep.setString(7, this.otherParty);
+            prep.setInt(1, userId);
+            prep.setDouble(2, amount);
+            prep.setInt(3, date);
+            prep.setInt(4, endDate);
+            prep.setInt(5, currencyType);
+            prep.setInt(6, frequency);
+            prep.setString(7, otherParty);
             prep.executeUpdate();
         } catch (SQLException e) {
             //e.printStackTrace();
@@ -46,33 +44,30 @@ public class Payment extends Transaction {
 
     public void update() {
         try {
-            Statement stmt = CryptoBudgetDatabase.connection.createStatement();
-            String update = "UPDATE PAYMENT SET USERID = ?, AMOUNT = ?, DATE = ?, ENDDATE = ?, CURRENCYTYPE = ?, " +
-                    "FREQUENCY = ?, OTHERPARTY = ? WHERE PAYMENTID = ?;";
+            String update = "UPDATE PAYMENT SET AMOUNT = ?, DATE = ?, ENDDATE = ?, CURRENCYTYPE = ?, " +
+                    "FREQUENCY = ?, OTHERPARTY = ? WHERE PAYMENTID = ? AND USERID = ?;";
             PreparedStatement prep = CryptoBudgetDatabase.connection.prepareStatement(update);
-            prep.setInt(1, userId);
-            prep.setDouble(2, amount);
-            prep.setInt(3, date);
-            prep.setInt(4, endDate);
-            prep.setInt(5, currencyType);
-            prep.setInt(6, frequency);
-            prep.setString(7, otherParty);
-            prep.setInt(8, id);
+            prep.setDouble(1, amount);
+            prep.setInt(2, date);
+            prep.setInt(3, endDate);
+            prep.setInt(4, currencyType);
+            prep.setInt(5, frequency);
+            prep.setString(6, otherParty);
+            prep.setInt(7, id);
+            prep.setInt(8, Main.currentUser.getUserId());
             prep.executeUpdate();
         } catch (SQLException e) {
             //e.printStackTrace();
         }
     }
 
-    // TODO only for user to logged in
     public static Payment[] getAllPayments() {
         try {
-            Statement stmt = CryptoBudgetDatabase.connection.createStatement();
-            String getAll = String.format("SELECT * FROM PAYMENT;");
-            ResultSet rs = stmt.executeQuery(getAll);
-
+            String getAll = String.format("SELECT * FROM PAYMENT WHERE USERID = ?;");
+            PreparedStatement prep = CryptoBudgetDatabase.connection.prepareStatement(getAll);
+            prep.setInt(1, Main.currentUser.getUserId());
+            ResultSet rs = prep.executeQuery(getAll);
             ArrayList<Payment> result = new ArrayList();
-
             while (rs.next()) {
                 int id = rs.getInt("PAYMENTID");
                 int userId = rs.getInt("USERID");
@@ -83,7 +78,6 @@ public class Payment extends Transaction {
                 int frequency = rs.getInt("FREQUENCY");
                 String otherParty = rs.getString("OTHERPARTY");
                 result.add(new Payment(id, userId, currencyType, amount, date, endDate, frequency, otherParty));
-
             }
             return result.toArray(new Payment[0]);
         } catch (SQLException e) {
@@ -92,15 +86,12 @@ public class Payment extends Transaction {
         return null;
     }
 
-    // TODO only for user logged in
     public static Payment getPayment(int findId) {
         try {
-            Statement stmt = CryptoBudgetDatabase.connection.createStatement();
-            String findPayment = "SELECT * FROM PAYMENT WHERE PAYMENTID = ?;";
+            String findPayment = "SELECT * FROM PAYMENT WHERE PAYMENTID = ? AND USERID = ?;";
             PreparedStatement prep = CryptoBudgetDatabase.connection.prepareStatement(findPayment);
-
             prep.setInt(1, findId);
-
+            prep.setInt(2, Main.currentUser.getUserId());
             ResultSet rs = prep.executeQuery();
             while (rs.next()) {
                 int id = rs.getInt("PAYMENTID");
@@ -111,7 +102,6 @@ public class Payment extends Transaction {
                 int currencyType = rs.getInt("CURRENCYTYPE");
                 int frequency = rs.getInt("FREQUENCY");
                 String otherParty = rs.getString("OTHERPARTY");
-
                 Payment result = new Payment(id, userId, currencyType, amount, date, endDate, frequency, otherParty);
                 return result;
             }
@@ -123,10 +113,10 @@ public class Payment extends Transaction {
 
     public static void removePayment(int removeId) {
         try {
-            Statement stmt = CryptoBudgetDatabase.connection.createStatement();
-            String remove = "DELETE FROM PAYMENT WHERE PAYMENTID = ?;";
+            String remove = "DELETE FROM PAYMENT WHERE PAYMENTID = ? AND USERID = ?;";
             PreparedStatement prep = CryptoBudgetDatabase.connection.prepareStatement(remove);
             prep.setInt(1, removeId);
+            prep.setInt(2, Main.currentUser.getUserId());
             prep.executeUpdate();
         } catch (SQLException e) {
             //e.printStackTrace();
